+++ src/components/MetadataSources/SourceManagement/SelectAllCollections.js
@@ -1,32 +1,33 @@
 import PropTypes from 'prop-types';
 import { useState } from 'react';
 import { FormattedMessage } from 'react-intl';
-import { useMutation } from 'react-query';
 
-import { useOkapiKy } from '@folio/stripes/core';
 import {
   Button,
   Col,
   Modal,
   Row,
 } from '@folio/stripes/components';
+import { useOkapiKyMutation } from '@folio/stripes-leipzig-components';
 
-import { COLLECTIONS_SELECT_ALL_BY_SOURCE_ID_API } from '../../../util/constants';
+import {
+  API_COLLECTIONS_SELECT_ALL_BY_SOURCE_ID,
+  QK_SOURCES,
+} from '../../../util/constants';
 
 const SelectAllCollections = ({
   sourceId,
   stripes,
 }) => {
-  const ky = useOkapiKy();
   const [showInfoModal, setShowInfoModal] = useState(false);
   const [modalText, setModalText] = useState('');
 
   const hasSelectAllCollectionsPerms = stripes.hasPerm('ui-finc-select.edit');
 
-  const { mutate: selectAllCollections } = useMutation({
-    mutationFn: async (id) => {
-      return ky.put(COLLECTIONS_SELECT_ALL_BY_SOURCE_ID_API(id), { json: { select: true } });
-    },
+  const { useUpdate } = useOkapiKyMutation({
+    queryKey: [QK_SOURCES, sourceId],
+    mutationKey: [API_COLLECTIONS_SELECT_ALL_BY_SOURCE_ID(sourceId)],
+    api: API_COLLECTIONS_SELECT_ALL_BY_SOURCE_ID(sourceId),
     onSuccess: () => {
       setModalText(<FormattedMessage id="ui-finc-select.source.modal.selectAllCollections.success" />);
       setShowInfoModal(true);
@@ -36,6 +37,8 @@ const SelectAllCollections = ({
       setShowInfoModal(true);
     }
   });
+
+  const { mutateAsync: selectAllCollections } = useUpdate();
 
   const handleClose = () => {
     setShowInfoModal(false);
@@ -49,7 +52,8 @@ const SelectAllCollections = ({
             buttonStyle="primary"
             disabled={!hasSelectAllCollectionsPerms}
             id="selectAllCollections"
-            onClick={() => selectAllCollections(sourceId)}
+            // onClick={() => selectAllCollections(sourceId)}
+            onClick={() => selectAllCollections({ select: true })}
           >
             <FormattedMessage id="ui-finc-select.source.button.selectAllCollections" />
           </Button>