+++ src/settings/CredentialsSettings.js
@@ -1,31 +1,32 @@
-import { useQuery, useMutation } from 'react-query';
 import { FormattedMessage } from 'react-intl';
 
-import { useOkapiKy } from '@folio/stripes/core';
 import { Loading, MessageBanner } from '@folio/stripes/components';
+import {
+  useOkapiKyMutation,
+  useOkapiKyQuery,
+} from '@folio/stripes-leipzig-components';
 
-import { EZB_CREDENTIALS_API } from '../util/constants';
+import {
+  API_EZB_CREDENTIALS,
+  QK_EZB_CREDENTIALS,
+} from '../util/constants';
 import CredentialsSettingsForm from './CredentialsSettingsForm';
 
 const CredentialsSettings = () => {
-  const ky = useOkapiKy();
-
-  const { data: credentials, error, isLoading, refetch } = useQuery({
-    queryKey: ['ezbCredentials'],
-    queryFn: async () => {
-      const res = await ky.get(EZB_CREDENTIALS_API).json();
-      return res ?? {};
-    }
+  const { data: credentials = {}, isError, isLoading, refetch } = useOkapiKyQuery({
+    queryKey: [QK_EZB_CREDENTIALS],
+    api: API_EZB_CREDENTIALS,
   });
 
-  const { mutate: updateCredentials } = useMutation({
-    mutationFn: async (values) => {
-      return ky.put(EZB_CREDENTIALS_API, { json: values });
-    },
+  const { useUpdate } = useOkapiKyMutation({
+    mutationKey: [QK_EZB_CREDENTIALS],
+    api: API_EZB_CREDENTIALS,
     onSuccess: () => {
       refetch();
     }
   });
+
+  const { mutateAsync: updateCredentials } = useUpdate();
 
   const handleSubmit = (values) => {
     updateCredentials(values);
@@ -35,7 +36,7 @@ const CredentialsSettings = () => {
     return <Loading />;
   }
 
-  if (error) {
+  if (isError) {
     return (
       <MessageBanner type="error">
         <FormattedMessage id="ui-finc-select.settings.ezbCredentials.error" />